requirements.txt
asgiref==3.5.2
certifi==2022.6.15
Django==4.0.5
django-haystack==3.2.1
elasticsearch==7.17.4
mysqlclient==2.1.0
Pillow==9.1.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
